Create Assign position and variable
The two players will start in place 0 of the board. We need to create two variables for each player, the name of the player and the second one to save the position on the board. From the second variable that we created for the position of the player on the board, we need to sum the random number that the player got after throwing the dice.